Move metrics_data to form model
Moves the metrics_data to form model, since it's a property of the form.
Updates the tests and removes the live_form controller spec file, since
it no longer has any tests in it.
