+++ app/views/live/show_form.html.erb
@@ -9,7 +9,7 @@
     </h1>
     <%= render FormStatusTagDescriptionComponent::View.new(status: :live) %>
 
-    <%= render MetricsSummaryComponent::View.new(form.made_live_date, metrics_data) %>
+    <%= render MetricsSummaryComponent::View.new(form.made_live_date, form.metrics_data) %>
 
     <h2 class="govuk-heading-l">Your form</h2>
